+++ public/assets/favicon.ico
