+++ public/index.html
@@ -10,7 +10,7 @@
     <link
       id="themeStylesheet"
       rel="stylesheet"
-      href="/styles/themes/default.css"
+      href="styles/themes/default.css"
     />
     <style>
       @import url("https://fonts.googleapis.com/css2?family=Inter:wght@300;400&family=JetBrains+Mono&family=Roboto&display=swap");
@@ -59,6 +59,6 @@
         <div class="text">Discord</div>
       </a>
     </div>
+    <script src="scripts/main.js"></script>
   </body>
 </html>
-<script src="main.js"></script>
