+++ public/rungeon.html
@@ -135,8 +135,6 @@
       </div>
       </div>
       </div>
-
-      <script defer src="main.js">
-    </script>
+      <script defer src="scripts/main.js"></script>
   </body>
 </html>
