+++ public/styles/themes/default.css
